% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/birthday.R
\name{birthday}
\alias{birthday}
\title{Birthday Problem Probability}
\usage{
birthday(x)
}
\arguments{
\item{x}{An integer or numeric vector giving the group size(s).}
}
\value{
A numeric vector of probabilities.
}
\description{
Computes the probability of at least two people in a group with size n sharing the same birthday.
}
\examples{
birthday(20:25)
}
