% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/firedam.R
\docType{data}
\name{firedam}
\alias{firedam}
\title{Fire Damage Data}
\format{
A data frame with X rows and Y variables:
\describe{
  \item{DAMAGE}{amount of damage taken}
  \item{DISTANCE}{distance from damage source at time of damage}
}
}
\source{
Original data file: firedam.rda
}
\usage{
data(firedam)
}
\description{
A dataset containing information on fire damage.
}
\keyword{datasets}
