% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lab5f.R
\name{lab5f}
\alias{lab5f}
\title{sample of hypergeometric function}
\usage{
lab5f(n = 1000, iter = 30, time = 1)
}
\arguments{
\item{n}{Number of samples per iteration}

\item{iter}{Number of iterations}

\item{time}{Pause between plots}
}
\value{
A series of barplots.
}
\description{
sample of hypergeometric function
}
