% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mybin.R
\name{mybin}
\alias{mybin}
\title{mybin}
\usage{
mybin(iter = 100, n = 10, p = 0.7)
}
\arguments{
\item{iter}{int variable, number of iterations run}

\item{n}{int variable, number of trials}

\item{p}{int variable, probability of "positive" outcome}
}
\value{
A bar plot
}
\description{
mybin
}
\examples{
mybin(100,10,0.7)
}
