% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myboot2.R
\name{myboot2}
\alias{myboot2}
\title{Title myboot2}
\usage{
myboot2(iter = 10000, x, fun = "mean", alpha = 0.05, cx = 1.5, ...)
}
\arguments{
\item{iter}{int, number of iterations of bootstrap}

\item{x}{sample for bootstrap}

\item{fun}{function for bootstrap}

\item{alpha}{alpha to determine confidence interval}

\item{cx}{scaling for text}

\item{...}{extra arguments for hist()}
}
\value{
bootstrap histogram, relevant quatities(invisable)
}
\description{
Title myboot2
}
\examples{
myboot2<-function(x,...)
}
