% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myci.R
\name{myci}
\alias{myci}
\title{Title myci}
\usage{
myci(x)
}
\arguments{
\item{x}{A vector representing the sample data}
}
\value{
A vector containing the lower and upper bounds of the 95 percent confidence interval
}
\description{
This function calculates a 95 percent confidence interval for the population mean
from a single sample.
}
\examples{
set.seed(23)
x <- rnorm(30, mean = 10, sd = 12)
myci(x)

}
