% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myclt.R
\name{myclt}
\alias{myclt}
\title{myclt}
\usage{
myclt(n, iter)
}
\arguments{
\item{n}{sample size}

\item{iter}{number of iterations}
}
\value{
a histogram of sample means.
}
\description{
myclt
}
\examples{
myclt(10, 10000)
}
