% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mycltb.R
\name{mycltb}
\alias{mycltb}
\title{Title Binom central limit theorem}
\usage{
mycltb(n, iter, p = 0.5, x = 0, ...)
}
\arguments{
\item{n}{int, sample size}

\item{iter}{int, number of samples taken}

\item{p}{float, probability of sucess}

\item{x}{monkey}

\item{...}{extra arguments}
}
\value{
a histogram
}
\description{
Title Binom central limit theorem
}
\examples{
mycltb(n=20,iter=10000,p=0.7,x=0)
}
