% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mymaxlik.R
\name{mymaxlik}
\alias{mymaxlik}
\title{Title mymaxlik}
\usage{
mymaxlik(lfun, x, param, ...)
}
\arguments{
\item{lfun}{The likelihood function to be maximized.}

\item{x}{A vector of values for the independent variable.}

\item{param}{A vector of parameter values to be optimized.}

\item{...}{extra arguments for the plotting function}
}
\value{
A list containing the index of the maximum likelihood estimate (MLE) (i),
the parameter values (parami) conesponding to the indexs, the log-likelihood at this MLE (yi),
and the slopes of the likelihood function around the MLE (slope).
}
\description{
Title mymaxlik
}
\examples{
logbin=function(x,param) log(dbinom(x,prob=param,size=10))
mymaxlik(x=c(9,9,1,9,9,9),param=seq(0,1,length=1000),lfun=logbin,xlab=expression(pi))
}
