% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myncurve.R
\name{myncurve}
\alias{myncurve}
\title{Title myncurve}
\usage{
myncurve(mu, sigma, a)
}
\arguments{
\item{mu}{integer for mu}

\item{sigma}{interger for sigma}

\item{a}{integer for a}
}
\value{
a graph with a filled in section, and the probability in a list
}
\description{
Title myncurve
}
\examples{
myncurve(mu=1, sigma=1, a=1)
}
