% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ntickets.R
\name{ntickets}
\alias{ntickets}
\title{Title ntickets}
\usage{
ntickets(N, gamma, p)
}
\arguments{
\item{N}{int, Number of Seats on the plane}

\item{gamma}{float, used to calculate confidence interval}

\item{p}{float, probability that an individual shows up}
}
\value{
plots showing the optimal amount of tickets to be sold for discrete and continuous, named list for nd,nc,N,p,and gamma
}
\description{
Title ntickets
}
\examples{
ntickets(N=400,gamma = 0.02, p = 0.95)
}
