% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myFunction.R
\name{predict_spruce}
\alias{predict_spruce}
\title{Predict spruce height from breast diameter}
\usage{
predict_spruce(x, coefs = NULL)
}
\arguments{
\item{x}{Numeric vector of breast height diameters in cm.}

\item{coefs}{Numeric vector of coefficients (default = NULL).
If NULL, the function will look for quad.lm in the global environment.}
}
\value{
Numeric vector of predicted heights in m.
}
\description{
This function uses coefficients from a quadratic regression model
to predict spruce tree height based on breast diameter.
}
\examples{
x <- c(10, 12, 14, 16, 18)
y <- 2 + 0.5 * x + 0.01 * x^2
quad.lm <- lm(y ~ poly(x, 2, raw = TRUE))
predict_spruce(c(15, 18, 20), coefs = coef(quad.lm))

}
