% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fire.R
\docType{data}
\name{fire}
\alias{fire}
\title{Fire data}
\format{
A data frame with X rows and Y variables:
\describe{
  \item{DAMAGE}{amount of damage taken}
  \item{DISTANCE}{distance from damage source at time of damage}
}
}
\source{
Original data file: fire.rda
}
\usage{
fire
}
\description{
A dataset containing fire information.
}
\keyword{datasets}
