% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myboot2.R
\name{myboot2}
\alias{myboot2}
\title{myboot2}
\usage{
myboot2(iter = 10000, x, fun = "mean", alpha = 0.05, cx = 1.5, ...)
}
\arguments{
\item{iter}{# of iterations}

\item{x}{sample vector}

\item{fun}{function used for calculation}

\item{alpha}{significance level}

\item{cx}{Character expansion factor for plot labels (optional)}

\item{...}{Additional arguments passed to the function}
}
\value{
A list containing the vector of bootstrap stats, the ci, observed value of the statistic for the original sample.
}
\description{
myboot2
}
\examples{
myboot2<-function(iter=10000,x,fun="median",alpha=0.05)
}
