% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myncurve.R
\name{myncurve}
\alias{myncurve}
\title{myncurve Function}
\usage{
myncurve(mu, sigma, a)
}
\arguments{
\item{mu}{The mean of the normal distribution.}

\item{sigma}{The standard deviation of the normal distribution.}

\item{a}{The starting x-value for shading.}
}
\value{
The probability between a and b.
}
\description{
This function draws a normal curve and shades the area
between two x-values. It also prints the probability value.
}
\examples{
myncurve(mu = 10, sigma = 4, a = 10)
}
