% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ntickets.R
\name{ntickets}
\alias{ntickets}
\title{Compute number of tickets to sell given seat count, show probability, and target overbook probability}
\usage{
ntickets(N, gamma, p)
}
\arguments{
\item{N}{integer number of seats on the plane}

\item{gamma}{probability that the plane is truly overbooked}

\item{p}{probability that a single purchaser shows up}
}
\value{
A named list with elements:
  \item{nd}{integer found using exact discrete binomial calculation}
  \item{nc}{integer found using normal approximation (continuity corrected)}
  \item{N}{input N}
  \item{p}{input p}
  \item{gamma}{input gamma}
}
\description{
Compute number of tickets to sell given seat count, show probability, and target overbook probability
}
\details{
The function creates two separate plots of the objective function:
The solution is the smallest integer n where the objective crosses zero.
}
\examples{
ntickets(N = 400, gamma = 0.02, p = 0.95)
}
